+++ migrations/20220102184512_make_md5_optional.sql
@@ -1,0 +1,12 @@
+ALTER TABLE roms
+RENAME COLUMN md5 TO md5_old;
+
+ALTER TABLE roms
+ADD COLUMN md5 VARCHAR(32);
+
+INSERT INTO roms (md5)
+SELECT md5_old
+FROM roms;
+
+ALTER TABLE roms
+DROP COLUMN md5_old;