+++ migrations/20220102213556_add_arcade_to_system.sql
@@ -1,0 +1,2 @@
+ALTER TABLE systems
+ADD COLUMN arcade BOOLEAN NOT NULL DEFAULT false;