+++ migrations/20220102215810_add_comment_to_games.sql
@@ -1,0 +1,2 @@
+ALTER TABLE games
+ADD COLUMN comment VARCHAR;