+++ migrations/20220103075227_add_merge_name_to_roms.sql
@@ -1,0 +1,2 @@
+ALTER TABLE roms
+ADD COLUMN merge_name VARCHAR;